[fix] Update imports after loss reorganization
Updated imports in files that reference moved losses:
- hybrid_pretrain_vicreg_loss.py: Update VICRegLoss import
- flickr_multistage_training.py: Update VICRegLoss import
- single_modality_strategy.py: Update VICRegLoss import

All imports now correctly reference the new locations:
- src.training.losses.self_supervised.VICRegLoss (was vicreg_loss)
- src.training.losses.wrappers.CombinedLoss (was combined_loss)

Verified all imports work correctly.
